PROYEC-1 Refactor de carpetas y nombres
