+++ src/main/java/com/eoi/springboot/ControladorUsuarios.java
@@ -1,0 +1,21 @@
+package com.eoi.springboot;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+
+public class ControladorUsuarios {
+
+
+    @RequestMapping("/")
+    String hellow(){
+        return "Hello world!";
+    }
+    @RequestMapping("/hola")
+    String hola(){
+        return "Hola mundo!";
+    }
+
+    @RequestMapping("/mimedicacion")
+    String mimedicacion(){
+        return "Tomate las pastillas!";
+    }
+}
