+++ src/main/java/com/eoi/springboot/MySpringBootAplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.eoi.springboot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
 @RestController
-public class DemoApplication {
+public class MySpringBootAplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(MySpringBootAplication.class, args);
 	}
