+++ src/test/java/com/eoi/springboot/MySpringBootAplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.demo;
+package com.eoi.springboot;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class MySpringBootAplicationTests {
 
 	@Test
 	void contextLoads() {
